avant dernier

+++ src/main/java/com/example/pi_ease/DAO/Repositories/CrediRepository.java
@@ -3,5 +3,5 @@ package com.example.pi_ease.DAO.Repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.pi_ease.DAO.Entities.Credit;
 
-public interface CrediRepository extends JpaRepository<Credit,Integer> {
+public interface CrediRepository extends JpaRepository<Credit,Long> {
 }
