+++ src/main/java/com/example/pi_ease/DAO/Repositories/TrancheRepository.java
@@ -3,5 +3,5 @@ package com.example.pi_ease.DAO.Repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.pi_ease.DAO.Entities.Tranche;
 
-public interface TrancheRepository extends JpaRepository<Tranche,Integer> {
+public interface TrancheRepository extends JpaRepository<Tranche, Long> {
 }
